v 0.30

+++ app/src/main/java/com/ashwini/weatherforecast/Model/Free/Clouds.kt
@@ -1,5 +1,5 @@
 package com.ashwini.weatherforecast.Model.Free
 
 data class Clouds(
-    val all:Int
+    val all:Double
 )
