+++ app/src/main/java/com/ashwini/weatherforecast/Model/Free/Wind.kt
@@ -6,5 +6,5 @@ data class Wind(
     @SerializedName("speed")
     val speed:Double,
     @SerializedName("deg")
-    val deg:Int
+    val deg:Double
     )
